Avoid rebuilding the property table when encoding a JsonSerializable
php_json_encode_array() takes care to avoid building the property
HashTable of an object being encoded. But when the object implements
JsonSerializable, a hashtable is built just so there is somewhere to put
the GC_PROTECTED flag.

So, protect the object instead. But a few tests (and probably users)
were doing var_dump($this) in jsonSerializable(), which fails due to
recursion if GC_PROTECTED is set on the object. So introduce a new
flag GC_PROTECTED2, allowing two things to recurse over the same object.